% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{braf_dnase_demo_dt_hg38}
\alias{braf_dnase_demo_dt_hg38}
\title{DNAaseI hypersensitivity sites for BRAF on hg38 as data.table}
\format{
\code{data.table}
}
\description{
DNAaseI hypersensitivity sites in the H562 cell line from UCSC Table Browser
hg38 liftover, subset to the BRAF locus with 1000 bp flank.
}
\keyword{data}
