% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{chromosome_arms_hg38}
\alias{chromosome_arms_hg38}
\title{Genomic coordinates of each chromosome arm on hg38 as GenomicRanges}
\format{
\code{GRanges}
}
\description{
A convenient single collection of the genomic span of each chromosome arm.
The span is described as:
|  ranges kept  |
p-arm         start |<------------->| centromere
q-arm    centromere |<------------->| end
}
\details{
Note, this data does not include a range across the centromeres, for these
see \code{exclusion_regions_hg38}.
}
\keyword{data}
