% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{cli_pio_colorscheme}
\alias{cli_pio_colorscheme}
\title{Set the color scheme options for pio CLI text in devgru}
\usage{
cli_pio_colorscheme()
}
\description{
Quickly set the options for pio package CLI text output.
}
\examples{
cli_pio_colorscheme()

}
