% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{cli_stopwatch_end}
\alias{cli_stopwatch_end}
\title{Verbose CLI output to designate a workflow end}
\usage{
cli_stopwatch_end(package = "devgru", function_name, stopwatch_start)
}
\arguments{
\item{package}{What is the package of the function, default: devgru}

\item{function_name}{What is the function name}

\item{stopwatch_start}{What was the start time of the stopwatch, typically captured
and passed along by \code{cli_stopwatch_start()}, see examples}
}
\description{
Display to the user the workflow has ended, stop the stopwatch, and display execution time.
}
\examples{
process_start <- cli_stopwatch_start(package = "devgru", function_name = "workflow_demo_func")
process_end <- cli_stopwatch_end(package = "devgru", function_name = "workflow_demo_func", stopwatch_start = process_start)

}
