% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{cli_stopwatch_start}
\alias{cli_stopwatch_start}
\title{Verbose CLI output to designate a workflow start}
\usage{
cli_stopwatch_start(package = "devgru", function_name)
}
\arguments{
\item{package}{What is the package of the function, default: devgru}

\item{function_name}{What is the function name}
}
\value{
CLI message and start time of process run
}
\description{
Display to the user the workflow has begun, the function that launched it, and
start the stopwatch to track execution time,
}
\examples{
process_start <- cli_stopwatch_start(package = "devgru", function_name = "workflow_demo_func")
process_start

}
\keyword{core}
