% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{copynumber_palettier}
\alias{copynumber_palettier}
\title{Create color palette for integer copy number states}
\usage{
copynumber_palettier(cn_states)
}
\arguments{
\item{cn_states}{Discrete integer values of copy number states, typically for plotting}
}
\value{
data.table of states and assigned color
}
\description{
Given a set of discrete integer copy number states, produce a diverging blue/red
color palette that scales to any number.
}
\examples{
# Most common set of states
copynumber_palettier(cn_states = seq(0,4))

# high level jump
copynumber_palettier(cn_states = c(1,2,3,8))

}
\keyword{core}
