% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{dt_sanitycheck}
\alias{dt_sanitycheck}
\title{Check if input is a data.table object, plus optional sanity check of column names}
\usage{
dt_sanitycheck(query_dt, expected_cols = NULL)
}
\arguments{
\item{query_dt}{Suspected data.table-like object to test}

\item{expected_cols}{Vector of character strings to check for consistency in
columns of query data.table, default: NULL}
}
\description{
Simple check if the input is a data.table object, plus includes an additional functionality
to perform a smart, transparent sanity check of expected column names.
}
\examples{
# Not a DT
dt_sanitycheck(query_dt = kras_dnase_demo_gr_hg38)
# Just a DT
dt_sanitycheck(query_dt = braf_dnase_demo_dt_hg38)
# DT with correct columns
dt_sanitycheck(query_dt = braf_dnase_demo_dt_hg38, expected_cols = c("seqnames","start","end","strand","signalValue","pValue","biospecimen","gene"))
# DT with less columns than expected
dt_sanitycheck(query_dt = braf_dnase_demo_dt_hg38[,-7], expected_cols = c("seqnames","start","end","strand","signalValue","pValue","biospecimen","gene"))

}
