% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{dt_to_gi}
\alias{dt_to_gi}
\title{Convert BEDPE-like data.table to GInteractions}
\usage{
dt_to_gi(input_dt, keep_metadata = TRUE)
}
\arguments{
\item{input_dt}{data.table with BEDPE-like columns}

\item{keep_metadata}{Logical, whether to preserve metadata (default: TRUE)}
}
\value{
GInteractions object
}
\description{
Single command to smartly convert a BEPDE-like data.table to a GIinteractions
object while preserving all metadata.
}
\keyword{core}
