% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{dt_to_gr}
\alias{dt_to_gr}
\title{Convert data.table to GRanges Object}
\usage{
dt_to_gr(input_dt)
}
\arguments{
\item{input_dt}{data.table object that minimally contains columns like chromosome start, and end position
that describe the genomic coordinates of the range}
}
\description{
Single command to smartly convert a data.table object to a GRanges object by
wrapping \code{gr_refactor_seqs()} around the \code{gUtils::dt2gr()} to avoid seqinfo conflicts
and apply proper sorting.
}
\examples{
# Converting from data.table to GRanges can be easily done with gUtils::dt2gr()
# however this single command conversion leaves some loose ends with the seqinfo

# Look at the original data.table
braf_dnase_demo_dt_hg38
# Now convert to GRanges
dt_to_gr(braf_dnase_demo_dt_hg38)
# Check the seqinfo
GenomeInfoDb::seqinfo(dt_to_gr(braf_dnase_demo_dt_hg38))

# Compared to the old way
# Conversion looks fine
gUtils::dt2gr(braf_dnase_demo_dt_hg38)
# However the seqinfo will conflict with other hg38 GRanges
GenomeInfoDb::seqinfo(gUtils::dt2gr(braf_dnase_demo_dt_hg38))

}
