% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{example_dnase_hg38}
\alias{example_dnase_hg38}
\title{DNAaseI hypersensitivity sites on hg38}
\format{
\code{GRanges}
}
\description{
DNAaseI hypersensitivity sites in the H562 cell line from UCSC Table Browser
hg19 that has been lifted over to hg38 using \code{rtracklayer::liftOver()}
}
\keyword{data}
