% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{exclusion_regions_hg38}
\alias{exclusion_regions_hg38}
\title{Collection of genomic regions to exclude from analysis on hg38 as GenomicRanges}
\format{
\code{GRanges}
}
\description{
A convenient single collection of genomic regions suggested to exclude from
analysis due to mappability issues, enrichment of duplicated regions,
telomeric, centromeric, acrocentric short-arms, and highly heterochromatin/gene
desert portions of the chromosomes.
}
\details{
Built from UCSC and ENCODE blacklists.
See excluderanges (https://github.com/dozmorovlab/excluderanges).
}
\keyword{data}
