% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{function_cli_intro}
\alias{function_cli_intro}
\title{Verbose CLI output for extended and rich description of steps along a
complex workflow, typically used in a pipeline}
\usage{
function_cli_intro(package = "devgru", function_name, ...)
}
\arguments{
\item{package}{What is the package of the function, default: devgru}

\item{function_name}{What is the function name}

\item{...}{What parameters were used in the function}
}
\description{
Display to the user the main package, the function being executed, and the parameters
passed to the function.
}
\examples{
# For demo purposes, set these variables
parameter_x <- 100
parameter_y <- "demo"

# In practice, these variables are set within the function being called as in below:
# workflow_demo_func <- function(parameter_x, parameter_y)
# then the function_cli_intro is called within the function

function_cli_intro(package = "devgru", function_name = "workflow_demo_func", parameter_x, parameter_y)

}
