% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{gene_body_hg38}
\alias{gene_body_hg38}
\title{Gene body segments from Ensemble v108 on hg38 as GenomicRanges}
\format{
\code{GRanges}
}
\description{
This genomic ranges contains the gene body segments (UTRs, Exons, Introns)
for a single-transcript representation of Ensembl v108 genes on hg38.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                     |  gene body schematic  |
}\if{html}{\out{</div>}}

[5\verb{ prime UTR]\{Exon\}<---Intron--->\{Exon\}<---Intron--->\{Exon\}[3} prime UTR]

The genes included had the following \code{gene_biotype}:
protein_coding
IG_C_gene,IG_D_gene,IG_J_gene,IG_V_gene,
TR_C_gene,TR_D_gene,TR_J_gene,TR_V_gene
lncRNA,miRNA

The transcript for protein-coding genes were determined using MANE Select &
MANE Plus Clinical v1.4 annotations

Built from
\itemize{
\item Ensembl v108 GTF (https://ftp.ensembl.org/pub/release-108/gtf/homo_sapiens/Homo_sapiens.GRCh38.108.gtf.gz)
\item MANE v1.4 summary file (https://ftp.ncbi.nlm.nih.gov/refseq/MANE/MANE_human/release_1.4/MANE.GRCh38.v1.4.summary.txt.gz)
}
}
\keyword{data}
