% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{geom_gap_imputation}
\alias{geom_gap_imputation}
\title{Diagnostic plot used for reviewing the imputation of gaps in segmentation}
\usage{
geom_gap_imputation(original_gap, imputed_gap_gr)
}
\arguments{
\item{original_gap}{GenomicRanges object of gap}

\item{imputed_gap_gr}{GenomicRanges object of imputed values across gap}
}
\value{
ggplot object
}
\description{
Creates a simple ggpplot2 style plot for quickly reviewing how the imputation of
each gap identified in a segmentation file

Note, this function was designed to be run as part of the \code{get_segmentation_gap_imputation()}
workflow.
}
\examples{
# Snippet from within the function `` that calls this geom
# geom_gap_imputation_diagnostic(
# original_gap = gap_of_interest,
# imputed_gap_gr = gUtils::grbind(partition_regression_data, imputed_sv_gaps))

}
\keyword{workflow}
