% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{geom_jabba_stats}
\alias{geom_jabba_stats}
\title{Diagnostic plot used for evaluating the performance of a JaBbA run}
\usage{
geom_jabba_stats(path_to_stats_dt, karyograph_dt)
}
\arguments{
\item{path_to_stats_dt}{Path to the \code{qc_stats.txt} output from \code{get_jabba_qc_diagnostic()} run}

\item{karyograph_dt}{Karyograph data.table from \code{JaBbA} run}
}
\value{
ggplot object
}
\description{
Creates a ggplot quilt-like plot for quickly evaluating the run performance
and output graph characterisics of a JaBbA run.

Note, this function was designed to be run as part of the \code{get_jabba_qc_diagnostic()}
workflow.
}
\examples{
# Snippet from within the function `get_jabba_qc_diagnostic` that calls this geom
# diagnostic_plot <- geom_jabba_stats(
# path_to_stats_dt = paste0(jabba_workdir_path, "qc_stats.txt"),
# karyograph_dt = data.table::data.table("cn" = kar$segstats$cn,
#                                        "cnmle" = kar$segstats$cnmle))

}
\keyword{workflow}
