% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_army_knife.R
\name{get_clonal_cnv_profile}
\alias{get_clonal_cnv_profile}
\title{Read in CNV profile DT of various flavors and extract a simplified clonal profile}
\usage{
get_clonal_cnv_profile(cnv_obj, caller, sample_id = NULL)
}
\arguments{
\item{cnv_obj}{CNV file in data.table or GRanges format}

\item{caller}{Name of caller that generated input CNV to be converted, supported: Battenberg and FACETS}

\item{sample_id}{Unique identifier to add to output, default: NULL}
}
\value{
ata.table object with rounded clonal CNV segments
}
\description{
Given a data.table or GRanges CNV object, extract the clonal CNV profile.
Any segment with a non-rounded value within 0.2 of the next integer value is rounded to that value.
The output will data.table will contain 6 columns: sample, seqnames, start, end, total, minor
The \code{sample} is either user-provided or row count placeholder
Currently supports CNV calls from Battenberg and FACETS
}
