% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_converter}
\alias{get_converter}
\title{Convert using appropriate format-specific functions}
\usage{
get_converter(input_obj, from_format, to_format)
}
\arguments{
\item{input_obj}{Full path to file}

\item{from_format}{Character string of input object format
"dt", "gr", "gi"}

\item{to_format}{Character string of input object format
"dt", "gr", "gi", "grl"}
}
\description{
Utilize the pre-built converters for various object formats.
Available:
\itemize{
\item dt --> gr
\item dt --> gi
\item gr --> dt
\item gi --> bedpe
\item gi --> grl
}
}
\keyword{internal}
