% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_cov_gap_imputation}
\alias{get_cov_gap_imputation}
\title{Generate gap imputed coverage from fragCounter output}
\usage{
get_cov_gap_imputation(
  path_to_fragcounter_cov,
  cpus = 1,
  make_diagnostic_plots = FALSE,
  path_to_diagnostic_plots_dir = NULL,
  exp_colnames = c("reads", "gc", "map", "reads.corrected"),
  verbose = T
)
}
\arguments{
\item{path_to_fragcounter_cov}{Path to fragCounter coverage RDS}

\item{cpus}{Number of CPUs to use for parallel execution, default: 1}

\item{make_diagnostic_plots}{Generate diagnostic plots for each imputed gap, default: FALSE}

\item{path_to_diagnostic_plots_dir}{Path to directory to place diagnostic plot PNGs}

\item{exp_colnames}{Expected column names for coverage input GR}

\item{verbose}{Output CLI during workflow, default: TRUE}
}
\value{
GenomicRanges object with gap imputed coverage
}
\description{
Given fragCounter coverage with small, zero signal gaps, run neighbrohood-based partition
regression gap imputation in a parallel manner to produce a gap imputed coverage file
that is provides cleaner foreground signal in JaBbA and GISTIC.
}
