% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_data}
\alias{get_data}
\title{Pulls in package data for use in package functions}
\usage{
get_data(name_of_data)
}
\arguments{
\item{name_of_data}{Character string with name of data as it appears in package}
}
\description{
Carbon copy of the the helper function described by stackoverflow user henfiber.
Pulls data into a temporary new environment and accesses it for assignment to
a variable in a function. For internal use only.
https://stackoverflow.com/questions/30951204/
}
\keyword{internal}
