% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_dryclean_segmentation}
\alias{get_dryclean_segmentation}
\title{Generate a segmentation file from a dryclean fragCounter read depth profile}
\usage{
get_dryclean_segmentation(
  path_to_dryclean_profile,
  threads = 1,
  random_seed = 999
)
}
\arguments{
\item{path_to_dryclean_profile}{Path to dryclean fragCounter read depth profile}

\item{threads}{Number of threads to use for parallel execution, default: 1}

\item{random_seed}{Numeric value to use as random seed for consistent execution}
}
\value{
data.table object with required GISTIC2.0 columns
}
\description{
Given dryclean processed fragCounter read depth profile, run DNAcopy's CBS algorithm
in a parallel manner to produce a segmentation file that is compatible with GISTIC2.0.
}
