% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_file_format}
\alias{get_file_format}
\title{Detect file format from filename and/or content}
\usage{
get_file_format(filename)
}
\arguments{
\item{filename}{Name of file with full path}
}
\description{
Detect a file's specific format, at scale
}
\keyword{internal}
