% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_fragcounter_segmentation}
\alias{get_fragcounter_segmentation}
\title{Generate a segmentation file from a fragCounter read depth profile}
\usage{
get_fragcounter_segmentation(
  path_to_fragcounter_profile,
  cpus = 1,
  random_seed = 999,
  verbose = T,
  exp_colnames = c("reads", "gc", "map", "reads.corrected")
)
}
\arguments{
\item{path_to_fragcounter_profile}{Path to fragCounter read depth profile, .rds file}

\item{cpus}{Number of CPUs to use for parallel execution, default: 1}

\item{random_seed}{Numeric value to use as random seed for consistent results}

\item{verbose}{Output CLI during workflow, default: TRUE}

\item{exp_colnames}{Expected column names for fragCounter read depth profile input}
}
\value{
data.table object with required GISTIC2.0 columns
}
\description{
Given fragCounter read depth profile, run DNAcopy's CBS algorithm in a parallel
manner to produce a segmentation file that is compatible with GISTIC2.0.
}
\examples{
# Demo fragCounter read depth profile
read_depth_demo_hg38

# Execute the full workflow
# fragcounter_seg <- get_fragcounter_segmentation(
# path_to_fragcounter_profile = "read_depth_demo_hg38.rds",
# cpus = 4)

}
\keyword{workflow}
