% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_imputed_gaps_per_chromosome}
\alias{get_imputed_gaps_per_chromosome}
\title{Single chromosome run imputation of small, zero coverage gaps in CBS segmentation}
\usage{
get_imputed_gaps_per_chromosome(
  chrom_for_imputation,
  gapless_segmentation_whitelist,
  segmentation_gaps,
  threshold = -4.95,
  structural_variant_breakpoints = NULL,
  sample_id,
  make_plots = F,
  path_for_plots
)
}
\arguments{
\item{chrom_for_imputation}{The chromosome to run gap imputation algorithm on}

\item{gapless_segmentation_whitelist}{GenomicRanges object of segmentation after removing blacklist regions and gaps}

\item{segmentation_gaps}{GenomicRanges object of gaps from the original segmentation}

\item{threshold}{Threshold of segmentation signal to determine a gap, default: -4.95}

\item{structural_variant_breakpoints}{GenomicRanges object of individual breakpoints from sample SVs}

\item{sample_id}{Name of sample}

\item{make_plots}{Generate diagnostic plots of imputation results? default: FALSE}

\item{path_for_plots}{Path to directory to hold diagnostic plot small PNGs (~80 KB per plot)}
}
\value{
data.table with gap imputed segmentation
}
\description{
Run imputation of small, zero coverage gaps in CBS segmentation using partition regression
of signal in neighborhood adjacent to the gap.
The gaps are typically small, less than 500 bp on average and inter spaced among large
segments of continuous, homogeneous signal. The gap signal is far less than expected
for real deletions, even homozygous events
Imputation greatly improves the foreground signal in downstream tools like JaBbA and GISTIC.

Note, this function was designed to be run as part of the \code{get_segmentation_gap_imputation()}
workflow. Also, see \code{chrompar()} for executing in parallel.
}
\keyword{workflow}
