% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_jabba_qc_diagnostic}
\alias{get_jabba_qc_diagnostic}
\title{Run diagnostic workflow of after JaBbA run to help evaluate quality of graph}
\usage{
get_jabba_qc_diagnostic(tumor_normal_id, jabba_workdir_path, verbose = T)
}
\arguments{
\item{tumor_normal_id}{The name of the tumor/normal pair with completed \code{JaBbA} run}

\item{jabba_workdir_path}{Path to the  completed \code{JaBbA} run work directory}

\item{verbose}{Output CLI during workflow, default: TRUE}
}
\value{
data.table of QC metrics and quilt-like ggplot
}
\description{
A simple rework to the JaBbA:::QCStats function for more specific local use cases
Provides a text file and quilt-like plot with metrics on convergence, RSME of copy number
estimates, purity, ploidy, loose ends, and junctions.
}
\examples{
# get_jabba_qc_diagnostic(
# tumor_normal_id = "CG15-2647-T_vs_190585-N",
# jabba_workdir_path = "jAbBa/CG15-2647-T_vs_190585-N_jabba_workdir/")

}
