% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_maf_lite}
\alias{get_maf_lite}
\title{Convert SNV & InDel mutation table to MAF-lite format}
\usage{
get_maf_lite(
  path_to_snv_dir,
  path_to_indel_dir,
  snv_consensus_filter = 2,
  indel_consensus_filter = 2,
  strict_samples = NULL,
  discovery_genes = NULL,
  return_type = "maf.lite"
)
}
\arguments{
\item{path_to_snv_dir}{Path to directory of MGP1000 union consensus SNVs}

\item{path_to_indel_dir}{Path to directory of MGP1000 union consensus InDels}

\item{snv_consensus_filter}{Threshold consensus filter for SNVs}

\item{indel_consensus_filter}{Threshold consensus filter for InDels}

\item{strict_samples}{Samples that will be filtered with \verb{*_consensus_filter``+1}}

\item{discovery_genes}{Gene set vector used for discovery with \verb{*_consensus_filter``=0}}

\item{return_type}{Output either the converted maf-lite format table or data.table of standard format mutation table, supported: maf.lite, data.table}
}
\value{
data.table like in either maf-lite or standard format
}
\description{
Read in multiple MGP1000 union consensus SNV & InDel mutation tables, filter based on consensus threshold
or sample/gene, merge all tables, calculate VAF, and prepare for downstream use in maf2maf or maf2vcf.
The bare minimum for a MAF is Chromosome, Start_Position, Reference_Allele, Tumor_Seq_Allele2, Tumor_Sample_Barcode
but will also include Matched_Norm_Sample_Barcode.
}
