% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_qc_diagnostics_snvindel}
\alias{get_qc_diagnostics_snvindel}
\title{Generate diagnostic plots for SNV & InDel variant calling QC checks}
\usage{
get_qc_diagnostics_snvindel(
  path_to_snv_dir = NULL,
  path_to_indel_dir = NULL,
  snv_indel_obj = NULL,
  plot_sample_names = T,
  include_caveman = F
)
}
\arguments{
\item{path_to_snv_dir}{Path to directory of MGP1000 union consensus SNVs}

\item{path_to_indel_dir}{Path to directory of MGP1000 union consensus InDels}

\item{snv_indel_obj}{Data.table object with merged SNVs & InDels, used in place of paths to directories}

\item{plot_sample_names}{Output plots to include sample names on y-axis, default: TRUE}

\item{include_caveman}{Add caveman to consensus list if used in SNV variant calling, default: FALSE}
}
\value{
Patchwork 'quilt'-like plot of ggplots
}
\description{
Single command to generate a comprehensive set of diagnostic plots that assist in
performing QC checks of SNVs & InDels called from MGP1000 using union consensus.
Can be used as both a first pass of unfiltered variants to see consensus skew as
well as used after filtering to show changes in calling metrics.
}
