% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_vaf}
\alias{get_vaf}
\title{Quick pull or explicitly calculate the VAF for mutation records of various flavors}
\usage{
get_vaf(vcf_obj, caller, mut_type)
}
\arguments{
\item{vcf_obj}{VCF file in data.table or GRanges format}

\item{caller}{Name of the caller that generated input VCF to be converted, supported: mutect, strelka, varscan, svaba, caveman}

\item{mut_type}{Type of mutations within VCF, supported: snv, indel}
}
\value{
data.table object with read support and VAF per record
}
\description{
Given a data.table or GRanges VCF object, quickly extract or explicitly calculate the VAF for all mutations.
For clarity, the read support for the VAF will be extracted as well.
Currently supports somatic SNV/InDel VCFs from Mutect, Strelka, Varscan, SvABA, CaVEMan.
}
