% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gi_refactor_seqs}
\alias{gi_refactor_seqs}
\title{Refactor seqinfo of GInteractions object for easy harmony}
\usage{
gi_refactor_seqs(input_gi, new_levels = gUtils::hg_seqlengths())
}
\arguments{
\item{input_gi}{GInteractions object to refactor}

\item{new_levels}{Named vector object used as the template for new seq details,
see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GInteractions object with updated seqinfo for both anchors
}
\description{
Single command to refactor all seq details of a GInteractions object to easily harmonize
with any other genomic object. This function updates seqinfo for both anchor regions.
By default, this package uses the autosome (1-22) and sex chromosomes (X,Y) of hg38,
see \code{gUtils::hg_seqlengths()}
Users can adjust this using the \code{new_levels} parameter.
}
\keyword{core}
