% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gi_to_dt}
\alias{gi_to_dt}
\title{Convert GInteractions to BEDPE data.table format}
\usage{
gi_to_dt(input_gi)
}
\arguments{
\item{input_gi}{GInteractions object}
}
\value{
data.table in BEDPE format
}
\description{
Converts a BEDPE representation GInteractions object into a BEDPE data.table while
preserving object metadata
}
\keyword{converter}
