% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gi_to_grl}
\alias{gi_to_grl}
\title{Convert GInteractions to GRangesList}
\usage{
gi_to_grl(input_gi, keep_metadata = TRUE)
}
\arguments{
\item{input_gi}{GInteractions object}

\item{keep_metadata}{Logical, whether to preserve metadata (default: TRUE)}
}
\value{
GRangesList where each element contains a pair of breakpoints
}
\description{
Single command to smartly convert GIinteractions to a GRangesList
object while preserving all metadata.
}
\keyword{core}
