% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gr_sanitycheck}
\alias{gr_sanitycheck}
\title{Check if input is a GenomicRanges object, plus optional sanity check of column names}
\usage{
gr_sanitycheck(query_gr, expected_cols = NULL)
}
\arguments{
\item{query_gr}{Suspected GenomicRanges-like object to test}

\item{expected_cols}{Vector of character strings to check for consistency in
columns of query GenomicRanges, default: NULL}
}
\value{
TRUE or FALSE
}
\description{
Simple check if the input is a GenomicRanges object, plus includes an additional functionality
to perform a smart, transparent sanity check of expected column names.
}
\examples{
# Not a GR
gr_sanitycheck(query_gr = braf_dnase_demo_dt_hg38)
# Just a GR
gr_sanitycheck(query_gr = kras_dnase_demo_gr_hg38)
# GR with correct columns
gr_sanitycheck(query_gr = kras_dnase_demo_gr_hg38,
               expected_cols = c("signalValue","pValue","biospecimen","gene"))
# GR with less columns than expected
gr_sanitycheck(query_gr = kras_dnase_demo_gr_hg38[,-3],
               expected_cols = c("signalValue","pValue","biospecimen","gene"))

}
