% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gr_to_seg}
\alias{gr_to_seg}
\title{Convert segmentation GenomicRanges object to a data.table with standardized columns and format}
\usage{
gr_to_seg(input_gr, exp_colnames = c("ID", "num.mark", "seg.mean"))
}
\arguments{
\item{input_gr}{Segmentation GenomicRanges object to convert to data.table}

\item{exp_colnames}{Expected column names for CBS segmentation input GenomicRanges}
}
\value{
data.table
}
\description{
Converts a segmentation GenomicRanges object into a data.table object then extracts the
appropriate columns and renames the columns to match the expected format.

The input GenomicRanges is typically generated from \code{get_dryclean_segmentation()} workflow
or other \code{DNAcopy} CBS algorithm runs.
}
\examples{
# Convert output from `get_dryclean_segmentation()` workflow
# dryclean_seg <- get_dryclean_segmentation(
# path_to_dryclean_profile = "read_depth_demo_hg38.rds",
# cpus = 4)

# gr_to_seg(dryclean_seg)

}
