% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{grl_refactor_seqs}
\alias{grl_refactor_seqs}
\title{Refactor seqinfo of GRangesList object for easy harmony}
\usage{
grl_refactor_seqs(input_grl, new_levels = gUtils::hg_seqlengths())
}
\arguments{
\item{input_grl}{GRangesList object to refactor}

\item{new_levels}{Named vector object used as the template for new seq details,
see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GRangesList object with updated seqinfo for all elements
}
\description{
Single command to refactor all seq details of a GRangesList object to easily harmonize
with any other genomic object. This function updates seqinfo for all GRanges elements
in the list.
By default, this package uses the autosome (1-22) and sex chromosomes (X,Y) of hg38,
see \code{gUtils::hg_seqlengths()}
Users can adjust this using the \code{new_levels} parameter.
}
\keyword{core}
