% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{grl_to_gi}
\alias{grl_to_gi}
\title{Convert GenomicRangesList to GInteractions}
\usage{
grl_to_gi(input_grl, keep_metadata = TRUE)
}
\arguments{
\item{input_grl}{GenomicRangesList where each element contains exactly 2 ranges}

\item{keep_metadata}{Logical, whether to preserve metadata (default: TRUE)}
}
\value{
GInteractions object
}
\description{
Converts a BEDPE representation GenomicRangesList object into a GInteractions object
while preserving object metadata
}
\keyword{converter}
