% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{hmcl_conditions_demo_dt_hg38}
\alias{hmcl_conditions_demo_dt_hg38}
\title{Condition labels for associated RNA-seq counts matrix as data.table}
\format{
\code{data.table}
}
\description{
Simple labels for RPMI-8226 and U-266 human myeloma cell lines that have
been processed to transcript counts matrix.
}
\keyword{data}
