% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{hmcl_counts_demo_dt_hg38}
\alias{hmcl_counts_demo_dt_hg38}
\title{RNA-seq counts matrix on hg38 as data.table}
\format{
\code{data.table}
}
\description{
Bulk RNA-seq data from RPMI-8226 and U-266 human myeloma cell lines that have
been processed to transcript counts matrix.
There are 3 replicates of each HMCL for use in DESeq2 differential gene
expression analysis.
}
\keyword{data}
