% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{kras_dnase_demo_gr_hg38}
\alias{kras_dnase_demo_gr_hg38}
\title{DNAaseI hypersensitivity sites for KRAS on hg38 as GenomicRanges}
\format{
\code{GRanges}
}
\description{
DNAaseI hypersensitivity sites in the H562 cell line from UCSC Table Browser
hg38 liftover, subset to the KRAS locus with 1000 bp flank.
}
\keyword{data}
