% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{myeloma_translocation_ig_loci_hg38}
\alias{myeloma_translocation_ig_loci_hg38}
\title{Ig loci of recurrent translocations in multiple myeloma as GenomicRanges}
\format{
\code{GenomicRanges}
}
\description{
This genomic ranges contains the known range of recurrent translocations at
the Ig loci (IGH/IGK/IGL) as well as a set with 500 Kb flank added.
}
\keyword{data}
