% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{myeloma_translocation_partner_loci_hg38}
\alias{myeloma_translocation_partner_loci_hg38}
\title{Partner loci of recurrent translocations in multiple myeloma as GenomicRanges}
\format{
\code{GenomicRanges}
}
\description{
This genomic ranges contains the known range of recurrent translocations distal
partner loci of the Ig translocations as well as a set with various flanks added.
}
\keyword{data}
