% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\docType{data}
\name{read_depth_demo_hg38}
\alias{read_depth_demo_hg38}
\title{dryclean-fragCounter read depth profile on hg38 as GenomicRanges}
\format{
\code{GRanges}
}
\description{
A slice of data from a \code{dryclean} adjusted \code{fragCounter} read depth profile
of a tumor sample.
The demo contains slices from chr19 and chr22.
}
\details{
See \code{fragCounter} (https://github.com/mskilab-org/fragCounter)
and \code{dryclean} (https://github.com/mskilab-org/dryclean)
}
\keyword{data}
