% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{sort_bedpe}
\alias{sort_bedpe}
\title{Sort a BEDPE}
\usage{
sort_bedpe(bedpe_dt)
}
\arguments{
\item{bedpe_dt}{data.table in BEDPE format}
}
\value{
Sorted data.table
}
\description{
Sort BEDPE data.table by genomic coordinates in chr-notation aware, header
adaptable manner
}
