% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_genes_shortcut}
\alias{get_genes_shortcut}
\title{Shortcut to get only protein coding genes from GTF file and convert to GRanges object}
\usage{
get_genes_shortcut(gtf_file_path, seq_lengths = gUtils::hg_seqlengths())
}
\arguments{
\item{gtf_file_path}{Path to GTF file}

\item{seq_lengths}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with GTF columns and updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Read in a GTF file, subset to protein coding genes, and convert it to a GRanges object with refactored seq details.
The GTF file can be either zipped or unzipped.
}
