% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_army_knife.R
\name{get_qc_diagnostics_alignment}
\alias{get_qc_diagnostics_alignment}
\title{Generate diagnostic plots for alignment QC checks using Alfred summary files}
\usage{
get_qc_diagnostics_alignment(
  path_to_tumor_dir = NULL,
  path_to_normal_dir = NULL,
  seq_protocol = "WGS"
)
}
\arguments{
\item{path_to_tumor_dir}{Path to directory of tumor sample Alfred summary files}

\item{path_to_normal_dir}{Path to directory of normal sample Alfred summary files}

\item{seq_protocol}{Type of sequencing protocol for display purposes, default: WGS}
}
\value{
Patchwork 'quilt'-like plot of ggplots
}
\description{
Single command to generate a comprehensive set of diagnostic plots that assist in
performing QC checks of tumor/normal read alignment of various sequencing protocol flavors
using Alfred summary files.
}
