% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{get_reader}
\alias{get_reader}
\title{Read file using appropriate format-specific reader}
\usage{
get_reader(
  file_path,
  file_format,
  delim,
  has_header,
  cpus,
  seq_lengths,
  verbose,
  ...
)
}
\arguments{
\item{file_path}{Full path to file}

\item{file_format}{Character string to force specific format detection: "auto" (default),
"bed", "bedpe", "maf", "vcf", "generic". Default: "auto"}

\item{delim}{Delimiter for generic files. Default: "auto"}

\item{has_header}{Logical, if files have header line. Default: TRUE}

\item{cpus}{Number of CPUs for reading data. Default: 1}

\item{seq_lengths}{Named vector for seqinfo refactoring (genomic formats only).
Default: gUtils::hg_seqlengths()}

\item{verbose}{Logical, print progress messages. Default: FALSE}

\item{...}{Additional arguments passed to format-specific readers}
}
\description{
Detect and utilize the necessary pre-built reader for each file of a specific
format. Otherwise, use generic data.table reader
}
\keyword{internal}
