% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gr_flank}
\alias{gr_flank}
\title{Precise addition of flanking to GenomicRanges with boundary-aware options}
\usage{
gr_flank(
  input_gr,
  start_flank = NULL,
  end_flank = NULL,
  start_flank_boundary = NULL,
  end_flank_boundary = NULL
)
}
\arguments{
\item{input_gr}{GenomicRanges object to add flanks to}

\item{start_flank}{Amount of flank (in bp) to be added to the start of the ranges, default: NULL}

\item{end_flank}{Amount of flank (in bp) to be added to the end of the ranges, default: NULL}

\item{start_flank_boundary}{The genomic coordinate the start of the flanked GenomicRanges should
to not exceed, default: NULL}

\item{end_flank_boundary}{The genomic coordinate the end of the flanked GenomicRanges should
to not exceed, default: NULL}
}
\value{
GenomicRanges object with updated start/end coordinates
}
\description{
Precise addition of flanks to the start and/or end of a GenomicRanges object with the option
of setting a boundary for the flanks, helpful to account for boundaries like chromosome arms.

Note, the boundaries are compared directly to the new flanked start/end and do not currently
take the chromosome into account automatically. Thus, this function is best applied in on
per chromosome basis.

If the flank exceeds the boundary, the new flanked start/end will set at the boundary.
}
\examples{
# The gene NOTCH2 is close to the p-arm centromere on chr1
notch2 <- gUtils::parse.gr("chr1:119911553-120069662")
notch2
chr1_p_arm <- gUtils::`\%Q\%`(chromosome_arms_hg38, seqnames == "chr1" & arm == "p")
chr1_p_arm
GenomicRanges::end(chr1_p_arm)

# If sampling data around this locus in a 3 Mb window, it's important to not
# extend into/beyond the centromere so we need to set this boundary
gr_flank(input_gr = notch2,
         start_flank = 3e6,
         end_flank = 3e6,
         start_flank_boundary = GenomicRanges::start(chr1_p_arm),
         end_flank_boundary = GenomicRanges::end(chr1_p_arm))

}
\keyword{core}
