% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{gr_refactor_seqs}
\alias{gr_refactor_seqs}
\title{Refactor seqinfo, seqnames, seqlengths, seqlevels of GRanges object for easy harmony}
\usage{
gr_refactor_seqs(input_gr, new_levels = gUtils::hg_seqlengths())
}
\arguments{
\item{input_gr}{GenomicRanges object to refactor}

\item{new_levels}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Single command to refactor all seq details of a GRanges object to easily harmonize with any other GRanges object.
By default, this package uses the autosome (1-22) and sex chromosomes (X,Y) of hg38, see \code{gUtils::hg_seqlengths()}
Users can adjust this using the \code{new_levels} parameter.
}
