% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{kit_loadout}
\alias{kit_loadout}
\title{Build the devgru environment by installing/loading packages}
\usage{
kit_loadout(update_kit = F)
}
\arguments{
\item{update_kit}{Update all packages even if already installed, default: TRUE}
}
\description{
Single command to install, if needed, and load all packages used in the devgru kit.
}
