% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read in a BED file, with or without header, and convert to GenomicRanges object}
\usage{
read_bed_file(
  bed_file_path,
  has_header = FALSE,
  additional_col_names = NULL,
  cpus = 1,
  seq_lengths = gUtils::hg_seqlengths(),
  allow_nonstandard_chr = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bed_file_path}{Path to BED file (supports .bed, .bed.gz)}

\item{has_header}{Logical, does BED file have header line. Default: FALSE}

\item{additional_col_names}{Character vector of names for additional columns in BED file,
beyond first three. Default: NULL}

\item{cpus}{Integer, number of CPUs for reading data, used by \code{data.table::fread()}. Default: 1}

\item{seq_lengths}{Named vector object used as template for new seq details.
See \code{gUtils::hg_seqlengths()} for example. Default: hg38 seqlengths}

\item{allow_nonstandard_chr}{Logical, whether to allow non-standard chromosomes
(decoys, HLA, alt contigs). Default: FALSE}

\item{verbose}{Logical, whether to print progress messages. Default: TRUE}
}
\value{
GenomicRanges object with BED columns, if present, and updated seqinfo, seqnames,
seqlengths, seqlevels
}
\description{
Read in a BED file and convert it to a GenomicRanges object with refactored seq details.
Expects the first 3 columns as chromosome, start, end; However column names are not necessary.
The BED file can be either zipped or unzipped. BED files use 0-based coordinates which are
automatically converted to 1-based for GenomicRanges.
}
\keyword{reader}
