% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_army_knife.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read in a BED file, with or without header, and convert to GRanges object}
\usage{
read_bed_file(
  bed_file_path,
  has_header = TRUE,
  col_names = NULL,
  seq_lengths = gUtils::hg_seqlengths()
)
}
\arguments{
\item{bed_file_path}{Path to BED file}

\item{has_header}{Indicate if files have a header line, expected to be same in all files}

\item{col_names}{Names for columns in BED file}

\item{seq_lengths}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with BED columns, if present, and updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Read in a BED file and convert it to a GRanges object with refactored seq details.
Expects the first 3 columns as chromosome, start, end; However column names are not necessary
The BED file can be either zipped or unzipped.
}
