% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_army_knife.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read in a BED file, with or without header, and convert to GRanges object}
\usage{
read_bed_file(
  bed_file_path,
  colnames = NULL,
  seq_lengths = gUtils::hg_seqlengths()
)
}
\arguments{
\item{bed_file_path}{Path to BED file}

\item{colnames}{Names for additional columns in BED file}

\item{seq_lengths}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with BED columns, if present, and updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Read in a BED file and convert it to a GRanges object with refactored seq details.
Expects the first 3 columns as chromosome, start, end; If more columns are present, user must provide column names.
The BED file must be unzipped.
}
