% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{read_gtf_file}
\alias{read_gtf_file}
\title{Read in a GTF file, such as one from Ensembl, and convert to GenomicRanges object}
\usage{
read_gtf_file(gtf_file_path, seq_lengths = gUtils::hg_seqlengths())
}
\arguments{
\item{gtf_file_path}{Path to GTF file}

\item{seq_lengths}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with GTF columns and updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Read in a GTF file which contains a number of columns and convert it to a GenomicRanges object with refactored seq details.
The GTF file can be either zipped or unzipped.
}
\keyword{reader}
