% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissArmyKnife.R
\name{read_maf_file}
\alias{read_maf_file}
\title{Read MAF file and convert to GenomicRanges object}
\usage{
read_maf_file(maf_file_path, cpus = 2, seq_lengths = gUtils::hg_seqlengths())
}
\arguments{
\item{maf_file_path}{Path to MAF file}

\item{cpus}{number of cpus for reading in data, used by \code{data.table::fread()}, default: 2}

\item{seq_lengths}{Named vector object used as the template for new seq details, see \code{gUtils::hg_seqlengths()} for example}
}
\value{
GenomicRanges object with MAF columns and updated seqinfo, seqnames, seqlengths, seqlevels
}
\description{
Read in a MAF file which contains a number of columns and convert it to a GenomicRanges object with refactored seq details.
The MAF file can be either zipped or unzipped.
For more specific MAFtools operations, see \code{maftools::read.maf()}
}
